api privacy + filter wip
