+++ .gitignore
@@ -14,6 +14,10 @@ dist-ssr
 coverage
 *.local
 
+# local env files
+.env.local
+.env.*.local
+
 /cypress/videos/
 /cypress/screenshots/
