+++ src/views/MovieCard.vue
@@ -1,38 +1,13 @@
 <script setup>
 import { ref } from "vue";
 import { RouterLink } from "vue-router";
-
-const showAll = ref(true)
-const showMovies = ref(false)
-const showSeries = ref(false)
-const showGames = ref(false)
-
-const handleShowMovies = () => {
-    showMovies.value = true;
-    showAll.value = false
-}
-const handleShowSeries = () => {
-    showSeries.value = true;
-    showAll.value = false
-}
-const handleShowGames = () => {
-    showGames.value = true;
-    showAll.value = false
-}
 defineProps({
     movieList: Array
 })
-
 </script>
 <template>
-    <div v-if="movieList.length != 0" class="movie-list">
-        <div class="sort-by">
-            <button @click="showAll = true" class="all">🍿 All</button>
-            <button @click="handleShowMovies" class="movies">🎬 Movies</button>
-            <button @click="handleShowSeries" class="series">🎞️ Series</button>
-            <button @click="handleShowGames" class="games">🎮 Games</button>
-        </div>
-        <ul v-if="movieList && showAll">
+    <div class="movie-list">
+        <ul>
             <li v-for="movie in movieList" :key="movie.imdbID">
                 <RouterLink :to="'/movies/' + movie.imdbID">
                     <img class="poster" :src="movie.Poster" :alt="movie.Title + ' Poster'" />
@@ -41,20 +16,10 @@ defineProps({
                 <h3>{{ movie.Title }}</h3>
             </li>
         </ul>
-        <ul v-if="movieList && showMovies">
-            <li v-for="movie in movieList" :key="movie.imdbID">
-                <RouterLink :to="'/movies/' + movie.imdbID">
-                    <img class="poster" :src="movie.Poster" :alt="movie.Title + ' Poster'" />
-                </RouterLink>
-                <div class="type">{{ movie.Type }}</div>
-                <h3>{{ movie.Title }}</h3>
-            </li>
-        </ul>
-        <h2 style="color: wheat;" v-if="movieList.length == 0">No results</h2>
     </div>
 </template>
 <style>
-button {
+button button {
     text-decoration: none;
     border: none;
     appearance: none;
@@ -105,58 +70,5 @@ li h4 {
     padding: 0.5rem;
     font-weight: 600;
     backdrop-filter: blur(5px);
-}
-
-.sort-by {
-    display: flex;
-    padding: 1rem;
-    justify-content: center;
-}
-.all {
-    color: #fff;
-    margin-right: 1rem;
-    background-color: #e07c24;
-    border-radius: 12px;
-    padding: 0.5rem;
-    border: solid 2px #e07c24;
-}
-.all:focus {
-    border: solid 2px red;
-}
-.movies {
-    color: #fff;
-
-    margin-right: 1rem;
-    background-color: #d82e2f;
-    border-radius: 12px;
-    padding: 0.5rem;
-    border: solid 2px #d82e2f;
-}
-.movies:focus {
-    border: 2px solid #e07c24;
-}
-.series {
-    color: #fff;
-
-    margin-right: 1rem;
-    background-color: #5a20cb;
-    border-radius: 12px;
-    padding: 0.5rem;
-    border: solid 2px #5a20cb;
-}
-.series:focus {
-    border: 2px solid #bf3312;
-}
-.games {
-    color: #fff;
-
-    margin-right: 1rem;
-    background-color: #8d3daf;
-    border-radius: 12px;
-    padding: 0.5rem;
-    border: solid 2px #8d3daf;
-}
-.games:focus {
-    border: 2px solid #bf3312;
 }
 </style>
